refactor: move modifier extensions to separate files
