+++ designSystem/src/main/java/com/example/designsystem/utils/modifierExtensions/DashedBorder.kt
@@ -1,26 +1,16 @@
-package com.example.designsystem.utils
+package com.example.designsystem.utils.modifierExtensions
 
 import android.annotation.SuppressLint
-import android.graphics.BlurMaskFilter
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.composed
 import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.draw.scale
 import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.geometry.Rect
 import androidx.compose.ui.geometry.RoundRect
-import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Paint
 import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.PathEffect
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.drawOutline
 import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
-import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 
 @SuppressLint("SuspiciousModifierThen")
@@ -66,42 +56,3 @@ fun Modifier.dashedBorder(
         )
     },
 )
-
-fun Modifier.dropShadow(
-    shape: Shape,
-    color: Color,
-    blur: Dp = 4.dp,
-    offsetY: Dp = 4.dp,
-    offsetX: Dp = 0.dp,
-    spread: Dp = 0.dp,
-) = this.drawBehind {
-    val shadowSize = Size(size.width + spread.toPx(), size.height + spread.toPx())
-    val shadowOutline = shape.createOutline(shadowSize, layoutDirection, this)
-
-    val paint = Paint()
-
-    paint.color = color
-
-    if (blur.toPx() > 0) {
-        paint.asFrameworkPaint().apply {
-            maskFilter = BlurMaskFilter(blur.toPx(), BlurMaskFilter.Blur.NORMAL)
-        }
-    }
-
-    drawIntoCanvas { canvas ->
-        canvas.save()
-        canvas.translate(offsetX.toPx(), offsetY.toPx())
-        canvas.drawOutline(shadowOutline, paint)
-        canvas.restore()
-    }
-}
-
-fun Modifier.mirroredContent(layoutDirection: LayoutDirection): Modifier {
-    val isRtl = layoutDirection == LayoutDirection.Rtl
-    return this.scale(scaleX = if (isRtl) -1f else 1f, scaleY = 1f)
-}
-
-fun Modifier.autoMirroredContent(): Modifier = composed {
-    val isRtl = LocalLayoutDirection.current == LayoutDirection.Rtl
-    this.scale(scaleX = if (isRtl) -1f else 1f, scaleY = 1f)
-}
